requirements.txt
py-cord[voice]
discord
youtube-dl
python-dotenv
youtube-search-python